Revert to the old testing suite and a few other bits.
This branch is specifically for debugging problems with Lily working
on octeon processors. It's unlikely to be useful for other purposes.

This commit includes a revert for two previous commits:

* Introducing the new test suite. The new test suite is Lily-driven,
  which is great when Lily is already up and running. In this case,
  the test harness itself can't even get to running. So this brings
  back the pre-commit-hook.py script which is Python-based. By being
  based in Python, multiple tests can run and give more output.

* Requiring scopes for class methods and vars. Since the old test
  suite was written without that, it was either revert that or fix the
  old tests that will go away when octeon is working. The choice was
  very easy. The octeon bugs reported were before the patch in
  question, so it shouldn't change the output.

A few other adjustments that have been made since the old test suite
was removed made some other tests fail. I removed a lot of tests that
I didn't feel like fixing. None of the tests removed were from the
octeon error log, so that shouldn't pose a problem.