Add packages dir with dis to root.
Use this if Lily seems to be generating bad bytecode. Hopefully it
won't come to needing this.

An easy way to use this is to drag the test that's failing into the
root directory. Add 'import dis' to the top of the test, and later
`print(dis.dis(__main__))`. That'll print out the disassembly and
shouldn't have any issues.