I think this is it: Fix the union order for lily_literal.
When Symtab creates a literal, it creates a lily_value which is then
cast over to lily_literal. Both lily_literal and lily_value share the
same fields and the same layout, so that value operations can transfer
the values over to the vm.

Or so I thought. Turns out that lily_literal had the 16 bit class id
first, and the flags second. lily_value puts the 32 bit flags first
and the 16 bit class id second.

When Symtab creates String and ByteString literals, it just so happens
to set the flags to remove the derefable marker. My theory is that
having a 16 bit value at the front of the struct (versus a 32 bit one)
causes the compiler/machine to see the struct differently, resulting
in the flag adjustment making the class id a totally wrong value.

As a precaution, I tested this change on the very latest master where
it worked flawlessly. Supposing that this fixes the ? issue, I'll
merge it into master.